Modal and HorizontalModal renamed to VerticalCohortSelector and HoritzontalCohortSelector, CalendarData typo fixed, className and Links updated on various pages
