+++ src/components/TestComponent/TestComponent.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import './TestComponent.css'
 import PurchaseBox from '../PurchaseBox/PurchaseBox';
-import Modal from "../Modal/Modal";
+import Modal from "../VerticalCohortSelector/VerticalCohortSelector";
 
 
 export default function TestComponent({ children }) {
