+++ src/pages/December.jsx
@@ -7,7 +7,7 @@ import Header from "../components/Header/Header";
 import GreenArticle from "../components/GreenArticle/GreenArticle";
 import Article from "../components/Article/Article";
 import Footer from "../components/Footer/Footer";
-import Modal from "../components/Modal/Modal";
+import Modal from "../components/VerticalCohortSelector/VerticalCohortSelector";
 import LeftContent from "../components/LeftContent/LeftContent";
 import TestComponent from "../components/TestComponent/TestComponent";
 import SpiffyCheckout from "../components/SpiffyCheckout/SpiffyCheckout";
