+++ src/pages/ScheduleDecember.jsx
@@ -7,7 +7,7 @@ import GreenArticle from "../components/GreenArticle/GreenArticle";
 import Footer from "../components/Footer/Footer";
 import TestComponent from "../components/TestComponent/TestComponent";
 import CalendarDec from "../components/Calendar/CalendarDec";
-import HorizontalModal from "../components/HorizontalModal/HorizontalModal";
+import HorizontalCohortSelector from "../components/HorizontalCohortSelector/HorizontalCohortSelector";
 import Navbar from "../components/Navbar/Navbar";
 
 export default function ScheduleDecember() {
@@ -26,7 +26,7 @@ export default function ScheduleDecember() {
                     <p>We welcome folks from other time zones to join us. </p>
                 </div>
             </TestComponent>
-            <HorizontalModal />
+            <HorizontalCohortSelector />
             < Footer />
         </div >
